Add pi0 training (#34)
This MR adds the training scripts + tutorial for the pi0 model.
Please see the training
[README.md](https://github.com/isaac-for-healthcare/i4h-workflows/tree/nigeln/pi0_training/workflows/robotic_ultrasound/scripts/training/pi_zero)
for more information.

Note:
The tutorial relies on the data collection from the state machine and
the playback script for validating data. Both of which are not yet in
this MR. Please comment if you believe these should be included in this
MR, or a subsequent one.

For testing use 5 previously collected state machine episodes
[here](https://drive.google.com/drive/folders/1CKnxRd2CRpoSGF-vB5SwavFVk6khTutZ?usp=sharing)
This is on our previous assets and it is only 5 episodes, so the
resulting model will not be very good :)

---------

Signed-off-by: Nigel Nelson <nigeln@nvidia.com>
Signed-off-by: Mingxin Zheng <mingxinz@nvidia.com>
Co-authored-by: Mingxin Zheng <18563433+mingxin-zheng@users.noreply.github.com>
Co-authored-by: Mingxin Zheng <mingxinz@nvidia.com>