Add State Machine + Data Collection (#44)
This MR adds the state machine which emulates a liver scan trajectory.
It includes the ability to collect these scans as training episodes for
IL policies by using the `.hdf5` format.

Please see the state machine
[README.md](https://github.com/isaac-for-healthcare/i4h-workflows/blob/f3a5ce5f5bfa4baf4fadf67bc0af6fc95a877bb9/workflows/robotic_ultrasound/scripts/simulation/README.md#liver-scan-state-machine)
for more comprehensive information.

Credit to @maximilianofir for the included updates that got the state
machine working with the latest USD files

---------

Signed-off-by: Nigel Nelson <nigeln@nvidia.com>
Signed-off-by: Nigel Nelson <61555377+NigelNelson@users.noreply.github.com>
Signed-off-by: Mingxin Zheng <mingxinz@nvidia.com>
Co-authored-by: Mingxin Zheng <mingxinz@nvidia.com>